Split IndexMergerTest and ImmutableConciseSetTest
