+++ processing/src/test/java/io/druid/segment/IndexIOTest.java
@@ -63,7 +63,7 @@ import java.util.Map;
 public class IndexIOTest
 {
   private static Interval DEFAULT_INTERVAL = Interval.parse("1970-01-01/2000-01-01");
-  private static final IndexSpec INDEX_SPEC = IndexMergerTest.makeIndexSpec(
+  private static final IndexSpec INDEX_SPEC = IndexMergerTestBase.makeIndexSpec(
       new ConciseBitmapSerdeFactory(),
       CompressedObjectStrategy.CompressionStrategy.LZ4,
       CompressedObjectStrategy.CompressionStrategy.LZ4,
