+++ processing/src/test/java/io/druid/segment/IndexMergerTestBase.java
@@ -50,10 +50,8 @@ import io.druid.segment.column.SimpleDictionaryEncodedColumn;
 import io.druid.segment.data.BitmapSerdeFactory;
 import io.druid.segment.data.CompressedObjectStrategy;
 import io.druid.segment.data.CompressionFactory;
-import io.druid.segment.data.ConciseBitmapSerdeFactory;
 import io.druid.segment.data.IncrementalIndexTest;
 import io.druid.segment.data.IndexedInts;
-import io.druid.segment.data.RoaringBitmapSerdeFactory;
 import io.druid.segment.incremental.IncrementalIndex;
 import io.druid.segment.incremental.IncrementalIndexAdapter;
 import io.druid.segment.incremental.IncrementalIndexSchema;
@@ -64,7 +62,6 @@ import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
 import javax.annotation.Nullable;
@@ -81,29 +78,21 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.Map;
 
-@RunWith(Parameterized.class)
-public class IndexMergerTest
+public class IndexMergerTestBase
 {
+  private final static IndexIO INDEX_IO = TestHelper.getTestIndexIO();
+
   @Rule
   public final TemporaryFolder temporaryFolder = new TemporaryFolder();
 
-  protected IndexMerger INDEX_MERGER;
-  private final static IndexIO INDEX_IO = TestHelper.getTestIndexIO();
+  protected IndexMerger indexMerger;
 
-  @Parameterized.Parameters(name = "{index}: useV9={0}, bitmap={1}, metric compression={2}, dimension compression={3}, long encoding={4}")
+  @Parameterized.Parameters(name = "{index}: metric compression={0}, dimension compression={1}, long encoding={2}")
   public static Collection<Object[]> data()
   {
     return Collections2.transform(
         Sets.cartesianProduct(
             ImmutableList.of(
-                ImmutableSet.of(
-                    true,
-                    false
-                ),
-                ImmutableSet.of(
-                    new RoaringBitmapSerdeFactory(null),
-                    new ConciseBitmapSerdeFactory()
-                ),
                 EnumSet.allOf(CompressedObjectStrategy.CompressionStrategy.class),
                 ImmutableSet.copyOf(CompressedObjectStrategy.CompressionStrategy.noNoneValues()),
                 EnumSet.allOf(CompressionFactory.LongEncodingStrategy.class)
@@ -143,8 +132,7 @@ public class IndexMergerTest
   @Rule
   public final CloserRule closer = new CloserRule(false);
 
-  public IndexMergerTest(
-      boolean useV9,
+  protected IndexMergerTestBase(
       BitmapSerdeFactory bitmapSerdeFactory,
       CompressedObjectStrategy.CompressionStrategy compressionStrategy,
       CompressedObjectStrategy.CompressionStrategy dimCompressionStrategy,
@@ -152,11 +140,6 @@ public class IndexMergerTest
   )
   {
     this.indexSpec = makeIndexSpec(bitmapSerdeFactory, compressionStrategy, dimCompressionStrategy, longEncodingStrategy);
-    if (useV9) {
-      INDEX_MERGER = TestHelper.getTestIndexMergerV9();
-    } else {
-      INDEX_MERGER = TestHelper.getTestIndexMerger();
-    }
   }
 
   @Test
@@ -170,7 +153,7 @@ public class IndexMergerTest
     final File tempDir = temporaryFolder.newFolder();
     QueryableIndex index = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist,
                 tempDir,
                 indexSpec
@@ -217,7 +200,7 @@ public class IndexMergerTest
     final File tempDir = temporaryFolder.newFolder();
     QueryableIndex index = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist,
                 tempDir,
                 indexSpec
@@ -259,7 +242,7 @@ public class IndexMergerTest
     final File tempDir = temporaryFolder.newFolder();
     QueryableIndex index = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist,
                 tempDir,
                 indexSpec
@@ -319,7 +302,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tempDir1,
                 indexSpec
@@ -333,7 +316,7 @@ public class IndexMergerTest
 
     QueryableIndex index2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist2,
                 tempDir2,
                 indexSpec
@@ -350,7 +333,7 @@ public class IndexMergerTest
     };
     QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(index1, index2),
                 true,
                 mergedAggregators,
@@ -408,7 +391,7 @@ public class IndexMergerTest
 
     final QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tmpDir1,
                 indexSpec
@@ -417,7 +400,7 @@ public class IndexMergerTest
     );
     final QueryableIndex index2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist2,
                 tmpDir2,
                 indexSpec
@@ -426,7 +409,7 @@ public class IndexMergerTest
     );
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(index1, index2),
                 true,
                 new AggregatorFactory[]{},
@@ -468,7 +451,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tempDir1,
                 indexSpec
@@ -488,7 +471,7 @@ public class IndexMergerTest
 
     QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 ImmutableList.of(index1),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -526,7 +509,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.append(
+            indexMerger.append(
                 ImmutableList.<IndexableAdapter>of(incrementalAdapter), null, tempDir1, indexSpec
             )
         )
@@ -547,7 +530,7 @@ public class IndexMergerTest
     AggregatorFactory[] mergedAggregators = new AggregatorFactory[]{new CountAggregatorFactory("count")};
     QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 ImmutableList.of(index1),
                 true,
                 mergedAggregators,
@@ -590,7 +573,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tempDir1,
                 indexSpec
@@ -623,7 +606,7 @@ public class IndexMergerTest
     AggregatorFactory[] mergedAggregators = new AggregatorFactory[]{new CountAggregatorFactory("count")};
     QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 ImmutableList.of(index1),
                 true,
                 mergedAggregators,
@@ -669,7 +652,7 @@ public class IndexMergerTest
     );
 
     QueryableIndex index1 = closer.closeLater(
-        INDEX_IO.loadIndex(INDEX_MERGER.persist(toPersist1, tempDir1, indexSpec))
+        INDEX_IO.loadIndex(indexMerger.persist(toPersist1, tempDir1, indexSpec))
     );
 
     final IndexableAdapter queryableAdapter = new QueryableIndexIndexableAdapter(index1);
@@ -683,7 +666,7 @@ public class IndexMergerTest
 
     QueryableIndex converted = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.convert(
+            indexMerger.convert(
                 tempDir1,
                 convertDir,
                 indexSpec
@@ -733,7 +716,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tempDir1,
                 indexSpec
@@ -766,7 +749,7 @@ public class IndexMergerTest
 
     QueryableIndex converted = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.convert(
+            indexMerger.convert(
                 tempDir1,
                 convertDir,
                 newSpec
@@ -840,7 +823,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tmpDir,
                 indexSpec
@@ -850,7 +833,7 @@ public class IndexMergerTest
 
     QueryableIndex index2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist2,
                 tmpDir2,
                 indexSpec
@@ -860,7 +843,7 @@ public class IndexMergerTest
 
     QueryableIndex index3 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist3,
                 tmpDir3,
                 indexSpec
@@ -871,7 +854,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(index1, index2, index3),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -948,7 +931,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tmpDir,
                 indexSpec
@@ -958,7 +941,7 @@ public class IndexMergerTest
 
     QueryableIndex index2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist2,
                 tmpDir2,
                 indexSpec
@@ -968,7 +951,7 @@ public class IndexMergerTest
 
     QueryableIndex index3 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist3,
                 tmpDir3,
                 indexSpec
@@ -978,7 +961,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(index1, index2, index3),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1029,7 +1012,7 @@ public class IndexMergerTest
 
     QueryableIndex indexA = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistA,
                 tmpDirA,
                 indexSpec
@@ -1039,7 +1022,7 @@ public class IndexMergerTest
 
     QueryableIndex indexB = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistB,
                 tmpDirB,
                 indexSpec
@@ -1049,7 +1032,7 @@ public class IndexMergerTest
 
     QueryableIndex indexB2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistB2,
                 tmpDirB2,
                 indexSpec
@@ -1059,7 +1042,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1071,7 +1054,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB2),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1197,7 +1180,7 @@ public class IndexMergerTest
 
     QueryableIndex indexA = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistA,
                 tmpDirA,
                 indexSpec
@@ -1207,7 +1190,7 @@ public class IndexMergerTest
 
     QueryableIndex indexB = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistB,
                 tmpDirB,
                 indexSpec
@@ -1217,7 +1200,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1339,7 +1322,7 @@ public class IndexMergerTest
 
     QueryableIndex indexA = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistA,
                 tmpDirA,
                 indexSpec
@@ -1349,7 +1332,7 @@ public class IndexMergerTest
 
     QueryableIndex indexB = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistB,
                 tmpDirB,
                 indexSpec
@@ -1359,7 +1342,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1480,7 +1463,7 @@ public class IndexMergerTest
 
     QueryableIndex indexA = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistA,
                 tmpDirA,
                 indexSpec
@@ -1490,7 +1473,7 @@ public class IndexMergerTest
 
     QueryableIndex indexB = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistB,
                 tmpDirB,
                 indexSpec
@@ -1500,7 +1483,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB),
                 false,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1590,7 +1573,7 @@ public class IndexMergerTest
 
     QueryableIndex indexA = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistA,
                 tmpDirA,
                 indexSpec
@@ -1600,7 +1583,7 @@ public class IndexMergerTest
 
     QueryableIndex indexB = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistB,
                 tmpDirB,
                 indexSpec
@@ -1610,7 +1593,7 @@ public class IndexMergerTest
 
     QueryableIndex indexBA = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistBA,
                 tmpDirBA,
                 indexSpec
@@ -1620,7 +1603,7 @@ public class IndexMergerTest
 
     QueryableIndex indexBA2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistBA2,
                 tmpDirBA2,
                 indexSpec
@@ -1630,7 +1613,7 @@ public class IndexMergerTest
 
     QueryableIndex indexC = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersistC,
                 tmpDirC,
                 indexSpec
@@ -1640,7 +1623,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB, indexBA, indexBA2),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1652,7 +1635,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(indexA, indexB, indexBA, indexC),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -1772,7 +1755,7 @@ public class IndexMergerTest
 
     final File tmpDirMerged = temporaryFolder.newFolder();
 
-    INDEX_MERGER.merge(
+    indexMerger.merge(
         toMerge,
         true,
         new AggregatorFactory[]{
@@ -1824,7 +1807,7 @@ public class IndexMergerTest
 
     final File tmpDirMerged = temporaryFolder.newFolder();
 
-    File merged = INDEX_MERGER.merge(
+    File merged = indexMerger.merge(
         toMerge,
         true,
         new AggregatorFactory[]{
@@ -1895,7 +1878,7 @@ public class IndexMergerTest
 
     final File tmpDirMerged = temporaryFolder.newFolder();
 
-    File merged = INDEX_MERGER.merge(
+    File merged = indexMerger.merge(
         toMerge,
         true,
         new AggregatorFactory[]{
@@ -1956,7 +1939,7 @@ public class IndexMergerTest
 
     final File tmpDirMerged = temporaryFolder.newFolder();
 
-    File merged = INDEX_MERGER.merge(
+    File merged = indexMerger.merge(
         toMerge,
         true,
         new AggregatorFactory[]{
@@ -2001,7 +1984,7 @@ public class IndexMergerTest
 
     final File tmpDirMerged = temporaryFolder.newFolder();
 
-    final File merged = INDEX_MERGER.merge(
+    final File merged = indexMerger.merge(
         toMerge,
         true,
         new AggregatorFactory[]{
@@ -2029,7 +2012,7 @@ public class IndexMergerTest
 
     QueryableIndex index1 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist1,
                 tmpDir,
                 indexSpec
@@ -2039,7 +2022,7 @@ public class IndexMergerTest
 
     QueryableIndex index2 = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 toPersist2,
                 tmpDir2,
                 indexSpec
@@ -2049,7 +2032,7 @@ public class IndexMergerTest
 
     final QueryableIndex merged = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.mergeQueryableIndex(
+            indexMerger.mergeQueryableIndex(
                 Arrays.asList(index1, index2),
                 true,
                 new AggregatorFactory[]{new CountAggregatorFactory("count")},
@@ -2160,7 +2143,7 @@ public class IndexMergerTest
     final File tempDir = temporaryFolder.newFolder();
     QueryableIndex index = closer.closeLater(
         INDEX_IO.loadIndex(
-            INDEX_MERGER.persist(
+            indexMerger.persist(
                 index1,
                 tempDir,
                 indexSpec
@@ -2301,7 +2284,7 @@ public class IndexMergerTest
     final File v9TmpDir = new File(tempDir, "v9-tmp");
 
     try {
-      INDEX_MERGER.persist(
+      indexMerger.persist(
           toPersist,
           tempDir,
           indexSpec
@@ -2423,6 +2406,6 @@ public class IndexMergerTest
     }
 
     final File tempDir = temporaryFolder.newFolder();
-    return closer.closeLater(INDEX_IO.loadIndex(INDEX_MERGER.persist(toPersist, tempDir, indexSpec)));
+    return closer.closeLater(INDEX_IO.loadIndex(indexMerger.persist(toPersist, tempDir, indexSpec)));
   }
 }
